Extract MetastoreValidator
